requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.0.1
numba==0.61.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
protobuf==5.29.5
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scikit-surprise==1.1.4
scipy==1.15.3
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
matplotlib==3.9.2
seaborn==0.13.2